remove mot metric
